+++ src/worker/queues/queues.ts
@@ -6,9 +6,9 @@ export const defaultJobOptions: JobsOptions = {
   attempts: 3,
   // Retries after 5s, 10s, 20s.
   backoff: { type: "exponential", delay: 5_000 },
-  // Purges completed jobs past 500 or 2 days.
+  // Purges completed jobs past 500 or 5 days.
   removeOnComplete: {
-    age: 60 * 60 * 24 * 2,
+    age: 60 * 60 * 24 * 5,
     count: 500,
   },
 };
@@ -33,7 +33,7 @@ export const logWorkerEvents = (worker: Worker) => {
       level: "error",
       message: `[${worker.name}] Failed: ${
         job?.id ?? "<no job ID>"
-      } with error: ${err.message} ${
+      } , Job Data: ${job?.data}, with error: ${err.message} ${
         env.NODE_ENV === "development" ? err.stack : ""
       }`,
       service: "worker",
